+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Toutes les modifications notables de ce projet seront documentées dans ce fichi
 ## refactory :
 - deplacement de l'impression dans un fichier à part
 
+# [2.6.1] - 
+## fix :
+- [#221](https://github.com/Refhi/Weda-Helper/issues/198) - l'affichage de l'historique à gauche ne fonctionnait plus dans certains cas.
+- [#227](https://github.com/Refhi/Weda-Helper/issues/227) - ajout des indicateurs de raccourcis dans Echanges Sécurisés
+- [#232](https://github.com/Refhi/Weda-Helper/issues/232) - correction de la détection des types de prescription d'ordo numérique
+- 
+
 # [2.6] - 
 ## ajouts :
 - [#162](https://github.com/Refhi/Weda-Helper/issues/162) - Mettre tout les raccourcis claviers lors d'un appuis long de Alt